+++ src/App.js
@@ -3,13 +3,12 @@ import Home from "./pages/Home";
 import SignIn from "./pages/SignIn";
 import SignUp from "./pages/SignUp";
 import ForgotPassword from "./pages/ForgotPassword";
-import FarmerProfil from "./pages/FarmerProfil";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import UploadProduct from "./pages/UploadProduct";
-import FarmerInventoryManager from "./pages/FarmerInventoryManager";
-import FarmerSignUp from "./pages/FarmerSignUp";
-import FarmerSignIn from "./pages/FarmerSignIn";
+import FarmerInventoryManager from "./pages/Farmer/FarmerInventoryManager";
+import FarmerSignUp from "./pages/Farmer/FarmerSignUp";
+import FarmerSignIn from "./pages/Farmer/FarmerSignIn";
 
 function App() {
   return (
@@ -20,7 +19,6 @@ function App() {
           <Route path="/sign-in" element={<SignIn />} />
           <Route path="/sign-up" element={<SignUp />} />
           <Route path="/forgot-password" element={<ForgotPassword />} />
-          <Route path="/farmer-profil" element={<FarmerProfil />} />
           <Route path="/upload-product" element={<UploadProduct />} />
           <Route
             path="/farmer-manage-inventory/:userId"
