+++ src/pages/Farmer/FarmerSignUp.jsx
@@ -1,20 +1,28 @@
 import { useState } from "react";
-import { assets } from "../assets";
+import { assets } from "../../assets";
 import { FaEye, FaEyeSlash } from "react-icons/fa";
-import { Link, useNavigate } from "react-router-dom";
-//import OAuth from "../components/OAuth";
-import { signInWithEmailAndPassword, getAuth } from "firebase/auth";
+// import OAuth from '../components/OAuth';
+import { doc, serverTimestamp, setDoc } from "firebase/firestore";
+import {
+  getAuth,
+  createUserWithEmailAndPassword,
+  updateProfile,
+} from "firebase/auth";
+import { db } from "../../config/firebase";
+import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
-import FarmerSignInUp from "../components/FarmerSignInUp";
+import FarmerSignInUp from "../../components/Farmer/FarmerSignInUp";
 
-function FarmerSignIn() {
+function FarmerSignUp() {
   const [showPassword, setShowPassword] = useState(false);
   const [formData, setFormData] = useState({
+    name: "",
     email: "",
     password: "",
+    address: "",
   });
 
-  const { email, password } = formData;
+  const { name, email, password, address } = formData;
 
   const navigate = useNavigate();
 
@@ -27,30 +35,37 @@ function FarmerSignIn() {
 
   async function onSubmit(e) {
     e.preventDefault();
+
     try {
       const auth = getAuth();
-      const userCredential = await signInWithEmailAndPassword(
+      const userCredential = await createUserWithEmailAndPassword(
         auth,
         email,
         password
       );
-      if (userCredential.user) {
-        const userId = auth.currentUser.uid;
-        navigate(`/farmer-manage-inventory/${userId}`);
-      }
+
+      updateProfile(auth.currentUser, {
+        displayName: name,
+      });
+      const user = userCredential.user;
+      const formDataCopy = { ...formData };
+      delete formDataCopy.password;
+      formDataCopy.timestamp = serverTimestamp();
+      await setDoc(doc(db, "users", user.uid), formDataCopy);
+      const userId = user.uid;
+      navigate(`/farmer-manage-inventory/${userId}`);
     } catch (error) {
-      console.log("sign", error);
-      toast.error("Bad user credentials");
+      console.log("error", error);
+      toast.error("Oops... Something went wrong with the registration", error);
     }
   }
 
   return (
     <section className="bg-primary">
-      <h1>Farmer SignIn</h1>
       <div className="flex justify-around items-center py-4 mb-8 max-w-[30rem] mx-auto">
         <h2 className="font-body text-5xl font-semibold p-4 text-font-light">
           Sign <br />
-          In
+          Up For Farmer
         </h2>
         <img
           className="w-[10rem] drop-shadow-md"
@@ -63,7 +78,23 @@ function FarmerSignIn() {
         <FarmerSignInUp />
         <div className="max-w-[30rem] mx-auto">
           <form onSubmit={onSubmit}>
-            <div className="mb-6">
+            <div className="mb-4">
+              <label
+                htmlFor="full name"
+                className="text-font-middle px-2 py-2 font-light"
+              >
+                Full Name
+              </label>
+              <input
+                className="font-body text-sm w-full my-2 px-4 py-2 text-font-middle 
+                rounded-3xl bg-[#F9F5F1] transition ease-in-out focus:border-[#FEFDFC]"
+                type="text"
+                id="name"
+                value={name}
+                onChange={onChange}
+              />
+            </div>
+            <div className="mb-4">
               <label
                 htmlFor="email"
                 className="text-font-middle px-2 py-2 font-light"
@@ -71,17 +102,31 @@ function FarmerSignIn() {
                 Email
               </label>
               <input
-                className="font-body text-sm w-full my-2 px-4 py-2 text-font-middle
+                className="font-body text-sm w-full my-2 px-4 py-2 text-font-middle 
                 rounded-3xl bg-[#F9F5F1] transition ease-in-out focus:border-[#FEFDFC]"
                 type="email"
                 id="email"
                 value={email}
                 onChange={onChange}
-                placeholder="email"
               />
             </div>
-
-            <div className="mb-6">
+            <div className="mb-4">
+              <label
+                htmlFor="address"
+                className="text-font-middle px-2 py-2 font-light"
+              >
+                Home Address
+              </label>
+              <input
+                className="font-body text-sm w-full my-2 px-4 py-2 text-font-middle 
+                rounded-3xl bg-[#F9F5F1] transition ease-in-out focus:border-[#FEFDFC]"
+                type="text"
+                id="address"
+                value={address}
+                onChange={onChange}
+              />
+            </div>
+            <div className="mb-4">
               <label
                 htmlFor="password"
                 className="text-font-middle px-2 py-2 font-light"
@@ -90,22 +135,21 @@ function FarmerSignIn() {
               </label>
               <div className="relative mb-6">
                 <input
-                  className="font-body text-sm w-full my-2 px-4 py-2 text-font-middle 
+                  className="mb-6 font-body text-sm w-full my-2 px-4 py-2 text-font-middle 
                 rounded-3xl bg-[#F9F5F1] transition ease-in-out focus:border-[#FEFDFC]"
                   type={showPassword ? "text" : "password"}
                   id="password"
                   value={password}
                   onChange={onChange}
-                  placeholder="password"
                 />
                 {showPassword ? (
                   <FaEyeSlash
-                    className="absolute right-4 top-4 text-xl cursor-pointer text-font-middle"
+                    className="absolute right-4 top-3.5 text-xl cursor-pointer text-font-middle"
                     onClick={() => setShowPassword((prevState) => !prevState)}
                   />
                 ) : (
                   <FaEye
-                    className="absolute right-4 top-4 text-xl cursor-pointer text-font-middle"
+                    className="absolute right-4 top-3.5 text-xl cursor-pointer text-font-middle"
                     onClick={() => setShowPassword((prevState) => !prevState)}
                   />
                 )}
@@ -113,20 +157,10 @@ function FarmerSignIn() {
             </div>
 
             <button type="submit" className="button-green mb-2">
-              Sign In
+              Sign Up
             </button>
-            <div>
-              <p className="text-center text-sm text-font-middle py-2">
-                <Link
-                  className="cursor-pointer underline"
-                  to="/forgot-password"
-                >
-                  Forgot password?
-                </Link>
-              </p>
-            </div>
             <div
-              className="my-4 max-w-xs m-auto flex items-center 
+              className="my-4 flex items-center max-w-xs m-auto
                   before:border-t 
                   before:flex-1 
                   before:text-font-middle
@@ -138,12 +172,7 @@ function FarmerSignIn() {
                 or sign in with
               </p>
             </div>
-            <div className="flex gap-12 justify-center py-4">
-              {/* Deleting this line as I don't see oAuth Component
-               <OAuth/>
-              <OAuth/>
-              <OAuth/> */}
-            </div>
+            <div className="flex gap-12 justify-center py-4"></div>
           </form>
         </div>
       </div>
@@ -151,4 +180,4 @@ function FarmerSignIn() {
   );
 }
 
-export default FarmerSignIn;
+export default FarmerSignUp;
