--- src/pages/FarmerProfil.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function FarmerProfil() {
-  return <div>FarmerProfil</div>;
-}
-
-export default FarmerProfil;
